Use Int32 instead of Single for Floor and Ceil

+++ Engine/Models/Map/MapSector.cs
@@ -4,8 +4,8 @@
     {
         public required int Id { get; set; }
         public List<Line> Walls { get; set; } = [];
-        public required float Floor { get; set; }
-        public required float Ceiling { get; set; }
+        public required int Floor { get; set; }
+        public required int Ceiling { get; set; }
         public required TextureInfo FloorTexture { get; set; }
         public required TextureInfo CeilingTexture { get; set; }
         public required short LightLevel { get; set; }
